Widen sessions.last_activity past 2038
The column holds seconds in 32 bits, so it stops in January 2038 — and Postgres
refuses the insert above 2^31 rather than waiting for the date, which turns a
clock skew into sessions that cannot be written. A new dialect test holds the
width against real Postgres and MySQL, since SQLite stores integers dynamically
and cannot fail this.

+++ packages/http/test/session-drivers.test.ts
@@ -10,7 +10,7 @@ beforeEach(async () => {
   await new SchemaBuilder(connection).create('sessions', (table) => {
     table.string('id').primary()
     table.text('payload')
-    table.integer('last_activity')
+    table.bigInteger('last_activity')
   })
 })
